FeedlyKit near complete, using 2-call method
FeedlyKit’s almost done as far as the scope of Dispatch is concerned.
This iteration uses a 2-call method to fetch articles — first calling
the Streams endpoint to get Entry IDs, then calling the Entries
endpoint with this IDs to get entries. Let’s consolidate that into the
one stream content GET request.